Control.Concurrent.STM.TVar

Description

TVars

data TVar a

Shared memory locations that support atomic memory transactions.

newTVar :: a -> STM (TVar a)

Create a new TVar holding a value supplied

newTVarIO :: a -> IO (TVar a)

IO version of newTVar. This is useful for creating top-level TVars using unsafePerformIO, because using atomically inside unsafePerformIO isn't possible.

readTVar :: TVar a -> STM a

Return the current value stored in a TVar.

readTVarIO :: TVar a -> IO a

Return the current value stored in a TVar. This is equivalent to

 readTVarIO = atomically . readTVar

but works much faster, because it doesn't perform a complete transaction, it just reads the current value of the TVar.

writeTVar :: TVar a -> a -> STM ()

Write the supplied value into a TVar.

modifyTVar :: TVar a -> (a -> a) -> STM ()

Mutate the contents of a TVar. N.B., this version is non-strict.

Since: stm-2.3

modifyTVar' :: TVar a -> (a -> a) -> STM ()

Strict version of modifyTVar.

Since: stm-2.3

stateTVar :: TVar s -> (s -> (a, s)) -> STM a

Like modifyTVar' but the function is a simple state transition that can return a side value which is passed on as the result of the STM.

Since: stm-2.5.0

swapTVar :: TVar a -> a -> STM a

Swap the contents of a TVar for a new value.

Since: stm-2.3

registerDelay :: Int -> IO (TVar Bool)

Switch the value of returned TVar from initial value False to True after a given number of microseconds. The caveats associated with threadDelay also apply.

Be careful not to exceed maxBound :: Int, which on 32-bit machines is only 2147483647 μs, less than 36 minutes.

mkWeakTVar :: TVar a -> IO () -> IO (Weak (TVar a))

Make a Weak pointer to a TVar, using the second argument as a finalizer to run when TVar is garbage-collected

Since: stm-2.4.3