Control.Monad.STM

Description

Software Transactional Memory: a modular composable concurrency abstraction. See

Composable memory transactions, by Tim Harris, Simon Marlow, Simon Peyton Jones, and Maurice Herlihy, in ACM Conference on Principles and Practice of Parallel Programming 2005. https://www.microsoft.com/en-us/research/publication/composable-memory-transactions/

This module only defines the STM monad; you probably want to import Control.Concurrent.STM (which exports Control.Monad.STM).

Note that invariant checking (namely the always and alwaysSucceeds functions) has been removed. See ticket #14324 and the removal proposal. Existing users are encouraged to encapsulate their STM operations in safe abstractions which can perform the invariant checking without help from the runtime system.

Documentation

data STM a

atomically :: STM a -> IO a

retry :: STM a

orElse :: STM a -> STM a -> STM a

check :: Bool -> STM ()

Check that the boolean condition is true and, if not, retry.

In other words, check b = unless b retry.

Since: stm-2.1.1

throwSTM :: Exception e => e -> STM a

catchSTM :: Exception e => STM a -> (e -> STM a) -> STM a

Orphan instances

MonadFix STM
Since: stm-2.3

Instance details

Methods

mfix :: (a -> STM a) -> STM a