Control.Concurrent.STM

Description

Software Transactional Memory: a modular composable concurrency abstraction. See

Composable memory transactions, by Tim Harris, Simon Marlow, Simon Peyton Jones, and Maurice Herlihy, in ACM Conference on Principles and Practice of Parallel Programming 2005. https://www.microsoft.com/en-us/research/publication/composable-memory-transactions/

Documentation

module Control.Monad.STM

module Control.Concurrent.STM.TVar

module Control.Concurrent.STM.TMVar

module Control.Concurrent.STM.TChan

module Control.Concurrent.STM.TQueue

module Control.Concurrent.STM.TBQueue

module Control.Concurrent.STM.TArray